refactor : alterando o tipo de dado que armazenará a matricula dos estudantes e será usado como id na tabela de banco de dados

+++ src/main/java/com/spring/apirestful/controllers/AlunosController.java
@@ -31,7 +31,7 @@ public class AlunosController {
     private AlunosService alunosService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<Alunos> findById (@PathVariable Long id) {
+    public ResponseEntity<Alunos> findById (@PathVariable Integer id) {
         Alunos aluno = this.alunosService.findById(id);
         return ResponseEntity.ok().body(aluno);
     }
